% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hsr_likert.r
\name{abbrev}
\alias{abbrev}
\title{Abberviate school names with exception substition}
\usage{
abbrev(x, exceptions = list(old = c("KAPS", "KCMS"), new = c("KAP", "KCCP")))
}
\arguments{
\item{x}{character vector of school names to abbreviatioe}

\item{exceptions}{list of two vectors of exceptions (character vector labeled \code{old}) and their
new abbreviations (character vector labeled \code{new})}
}
\description{
`abbrev` abbreviates
}

