% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hsr_likert.r
\name{pre_process_likert}
\alias{pre_process_likert}
\title{Preprocess HSR data for plotting Likert data}
\usage{
pre_process_likert(.data, school_order = c("KAP", "KAMS", "KCCP", "KBCP",
  "KC", "KN"), ...)
}
\arguments{
\item{.data}{\code{data.frame} of Healthy Schools and Regions data in long format}

\item{school_order}{order schools (abbreviated names) should be listed in}

\item{...}{args passed to \code{\link{abbrev}}.}
}
\description{
`plot_likert` generates a floating bar graph (horizontally) aligned
for survey data that is on lickert scale. Currently this function supports a a five
point Likert scale the with levels "Strongly Disagree", "Disagree", "Neutral",
"Agree", and "Strongly Agree"
}

